+++ lib/patron/header_parser.rb
@@ -1,7 +1,6 @@
-require 'strscan'
-
 module Patron::HeaderParser
-  CRLF = /#{Regexp.escape("\r\n")}/
+  HTTP_STATUS_LINE_START_RE = /^HTTP\/\d\.\d \d+/
+  HEADER_LINE_START_RE = /^[^:]+\:/
 
   # Returned for each response parsed out
   class SingleResponseHeaders < Struct.new(:status_line, :headers)
@@ -16,17 +15,14 @@ module Patron::HeaderParser
   #
   # @param [String] the string of headers, with responses delimited by empty lines. All lines must end with CRLF
   # @return Array<SingleResponseHeaders>
-  def self.parse(headers_from_multiple_responses_in_sequence)
-    s = StringScanner.new(headers_from_multiple_responses_in_sequence)
+  def self.parse(string_of_headers_from_multiple_responses_in_sequence)
     responses = []
-    until s.eos?
-      return unless scanned = s.scan_until(CRLF)
-      matched_line = scanned[0..-3]
-      if matched_line =~ /^HTTP\/\d\.\d \d+/
+    string_of_headers_from_multiple_responses_in_sequence.each_line do |matched_line|
+      if matched_line =~ HTTP_STATUS_LINE_START_RE
         responses << SingleResponseHeaders.new(matched_line.strip, [])
-      elsif matched_line =~ /^[^:]+\:/
+      elsif matched_line =~ HEADER_LINE_START_RE
         raise "Header should follow an HTTP status line" unless responses.any?
-        responses[-1].headers << matched_line
+        responses[-1].headers << matched_line.strip
       end # else it is the end of the headers for the request
     end
     responses